- Data routing and peer discovery mechanism has been updated - Unused code has been deleted - Update android plugin

+++ viper/src/main/java/com/w3engineers/ext/viper/application/data/remote/model/MeshAcknowledgement.java
@@ -12,10 +12,10 @@ import android.os.Parcelable;
  */
 public class MeshAcknowledgement implements Parcelable {
 
-    public int id;
+    public long id;
     public MeshPeer mMeshPeer;
 
-    public MeshAcknowledgement(int id) {
+    public MeshAcknowledgement(long id) {
         this.id = id;
     }
 
@@ -26,12 +26,12 @@ public class MeshAcknowledgement implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(this.id);
+        dest.writeLong(this.id);
         dest.writeParcelable(this.mMeshPeer, flags);
     }
 
     protected MeshAcknowledgement(Parcel in) {
-        this.id = in.readInt();
+        this.id = in.readLong();
         this.mMeshPeer = in.readParcelable(MeshPeer.class.getClassLoader());
     }
